+++ R/lmtp_Task.R
@@ -16,8 +16,10 @@ lmtp_Task <- R6::R6Class(
     bounds = NULL,
     folds = NULL,
     weights = NULL,
+    multivariate = NULL,
     initialize = function(data, trt, outcome, time_vary, baseline, cens, k,
-                          shift, shifted, id, outcome_type = NULL, V = 10, weights = NULL, bounds = NULL) {
+                          shift, shifted, id, outcome_type = NULL, V = 10,
+                          weights = NULL, bounds = NULL, bound = NULL) {
       self$tau <- determine_tau(outcome, trt)
       self$n <- nrow(data)
       self$trt <- trt
@@ -28,8 +30,9 @@ lmtp_Task <- R6::R6Class(
       self$survival <- outcome_type == "survival"
       self$bounds <- y_bounds(data[[final_outcome(outcome)]], self$outcome_type, bounds)
       data$lmtp_id <- create_ids(data, id)
-      self$id <- id
+      self$id <- data$lmtp_id
       self$folds <- setup_cv(data, V, data$lmtp_id, final_outcome(outcome), self$outcome_type)
+      self$multivariate <- is.list(trt)
 
       shifted <- {
         if (is.null(shifted) && !is.null(shift))