+++ R/sdr.R
@@ -1,19 +1,15 @@
-cf_sdr <- function(task, outcome, ratios, learners, control, progress_bar) {
-  out <- list()
-  for (fold in seq_along(task$folds)) {
+cf_sdr <- function(Task, outcome, ratios, learners, control, pb) {
+  out <- vector("list", length = length(Task$folds))
+  for (fold in seq_along(Task$folds)) {
     out[[fold]] <- future::future({
-      estimate_sdr(get_folded_data(task$natural, task$folds, fold),
-                   get_folded_data(task$shifted[, task$trt, drop = F], task$folds, fold),
-                   outcome, task$node_list$outcome,
-                   task$cens,
-                   task$risk,
-                   task$id,
-                   task$tau,
-                   task$outcome_type,
-                   get_folded_data(ratios, task$folds, fold)$train,
-                   learners,
-                   control,
-                   progress_bar)
+      estimate_sdr(
+        get_folded_data(Task$natural, Task$folds, fold),
+        get_folded_data(Task$shifted[, unlist(Task$trt), drop = F], Task$folds, fold),
+        Task$trt, outcome, Task$node_list$outcome,
+        Task$cens, Task$risk, Task$tau, Task$outcome_type,
+        get_folded_data(ratios, Task$folds, fold)$train,
+        learners, control, pb
+      )
     },
     seed = TRUE)
   }
@@ -25,8 +21,8 @@ cf_sdr <- function(task, outcome, ratios, learners, control, progress_bar) {
        fits = lapply(out, function(x) x[["fits"]]))
 }
 
-estimate_sdr <- function(natural, shifted, outcome, node_list, cens, risk, id, tau,
-                         outcome_type, ratios, learners, control, progress_bar) {
+estimate_sdr <- function(natural, shifted, trt, outcome, node_list, cens, risk, tau,
+                         outcome_type, ratios, learners, control, pb) {
 
   m_natural_train <- m_shifted_train <- cbind(matrix(nrow = nrow(natural$train),
                                                      ncol = tau),
@@ -36,7 +32,7 @@ estimate_sdr <- function(natural, shifted, outcome, node_list, cens, risk, id, t
                                                      ncol = tau),
                                               natural$valid[[outcome]])
 
-  fits <- list()
+  fits <- vector("list", length = tau)
 
   for (t in tau:1) {
     i  <- censored(natural$train, cens, t)$i
@@ -49,14 +45,18 @@ estimate_sdr <- function(natural, shifted, outcome, node_list, cens, risk, id, t
     vars <- node_list[[t]]
 
     if (t == tau) {
-      learners <- check_variation(natural$train[i & rt, ][[outcome]], learners)
-
-      fit <- run_ensemble(natural$train[i & rt, c(id, vars, outcome)],
+      fit <- run_ensemble(natural$train[i & rt, c("lmtp_id", vars, outcome)],
                           outcome,
                           learners,
                           outcome_type,
-                          id,
+                          "lmtp_id",
                           control$.learners_outcome_folds)
+
+      if (control$.return_full_fits) {
+        fits[[t]] <- fit
+      } else {
+        fits[[t]] <- extract_sl_weights(fit)
+      }
     }
 
     if (t < tau) {
@@ -68,32 +68,35 @@ estimate_sdr <- function(natural, shifted, outcome, node_list, cens, risk, id, t
 
       natural$train[, pseudo] <- shifted$train[, pseudo] <- densratio
 
-      learners <- check_variation(natural$train[i & rt, ][[pseudo]], learners)
-
-      fit <- run_ensemble(natural$train[i & rt, c(id, vars, pseudo)],
+      fit <- run_ensemble(natural$train[i & rt, c("lmtp_id", vars, pseudo)],
                           pseudo,
                           learners,
                           "continuous",
-                          id,
+                          "lmtp_id",
                           control$.learners_outcome_folds)
+
+      if (control$.return_full_fits) {
+        fits[[t]] <- fit
+      } else {
+        fits[[t]] <- extract_sl_weights(fit)
+      }
     }
 
-    if (control$.return_full_fits) {
-      fits[[t]] <- fit
+    if (length(trt) > 1) {
+      trt_t <- trt[[t]]
     } else {
-      fits[[t]] <- extract_sl_weights(fit)
+      trt_t <- trt[[1]]
     }
 
-    trt_var <- names(shifted$train)[t]
-    under_shift_train <- natural$train[jt & rt, vars]
-    under_shift_train[[trt_var]] <- shifted$train[jt & rt, trt_var]
+    under_shift_train <- natural$train[jt & rt, c("lmtp_id", vars)]
+    under_shift_train[, trt_t] <- shifted$train[jt & rt, trt_t]
 
-    under_shift_valid <- natural$valid[jv & rv, vars]
-    under_shift_valid[[trt_var]] <- shifted$valid[jv & rv, trt_var]
+    under_shift_valid <- natural$valid[jv & rv, c("lmtp_id", vars)]
+    under_shift_valid[, trt_t] <- shifted$valid[jv & rv, trt_t]
 
-    m_natural_train[jt & rt, t] <- bound(SL_predict(fit, natural$train[jt & rt, vars]), 1e-05)
+    m_natural_train[jt & rt, t] <- bound(SL_predict(fit, natural$train[jt & rt, c("lmtp_id", vars)]), 1e-05)
     m_shifted_train[jt & rt, t] <- bound(SL_predict(fit, under_shift_train), 1e-05)
-    m_natural_valid[jv & rv, t] <- bound(SL_predict(fit, natural$valid[jv & rv, vars]), 1e-05)
+    m_natural_valid[jv & rv, t] <- bound(SL_predict(fit, natural$valid[jv & rv, c("lmtp_id", vars)]), 1e-05)
     m_shifted_valid[jv & rv, t] <- bound(SL_predict(fit, under_shift_valid), 1e-05)
 
     m_natural_train[!rt, t] <- 0