+++ R/tmle.R
@@ -1,6 +1,6 @@
-cf_tmle <- function(task, outcome, cumulated, ratios, learners, control, progress_bar) {
-  out <- list()
-
+cf_tmle <- function(Task, outcome, cumulated, ratios, learners, control, pb) {
+  out <- vector("list", length = length(Task$folds))
+  
   if(cumulated == FALSE) {
     ratios <- matrix(t(apply(ratios, 1, cumprod)),
                      nrow = nrow(ratios),
@@ -9,20 +9,14 @@ cf_tmle <- function(task, outcome, cumulated, ratios, learners, control, progres
 
   for (fold in seq_along(task$folds)) {
     out[[fold]] <- future::future({
-      estimate_tmle(get_folded_data(task$natural, task$folds, fold),
-                    get_folded_data(task$shifted[, task$trt, drop = F], task$folds, fold),
-                    outcome,
-                    task$node_list$outcome,
-                    task$cens,
-                    task$risk,
-                    task$id,
-                    task$tau,
-                    task$outcome_type,
-                    get_folded_data(ratios, task$folds, fold)$train,
-                    task$weights[task$folds[[fold]]$training_set],
-                    learners,
-                    control,
-                    progress_bar)
+      estimate_tmle(
+        get_folded_data(Task$natural, Task$folds, fold),
+        get_folded_data(Task$shifted[, unlist(Task$trt), drop = F], Task$folds, fold),
+        Task$trt, outcome, Task$node_list$outcome, Task$cens, Task$risk,
+        Task$tau, Task$outcome_type,
+        get_folded_data(ratios, Task$folds, fold)$train,
+        learners, control, pb
+      )
     },
     seed = TRUE)
   }
@@ -35,12 +29,12 @@ cf_tmle <- function(task, outcome, cumulated, ratios, learners, control, progres
        fits = lapply(out, function(x) x[["fits"]]))
 }
 
-estimate_tmle <- function(natural, shifted, outcome, node_list, cens, risk, id,
-                          tau, outcome_type, ratios, weights, learners, control, progress_bar) {
+estimate_tmle <- function(natural, shifted, trt, outcome, node_list, cens,
+                          risk, tau, outcome_type, ratios, learners, control, pb) {
   m_natural_train <- m_shifted_train <- matrix(nrow = nrow(natural$train), ncol = tau)
   m_natural_valid <- m_shifted_valid <- matrix(nrow = nrow(natural$valid), ncol = tau)
 
-  fits <- list()
+  fits <- vector("list", length = tau)
   for (t in tau:1) {
     i  <- censored(natural$train, cens, t)$i
     jt <- censored(natural$train, cens, t)$j
@@ -56,13 +50,11 @@ estimate_tmle <- function(natural, shifted, outcome, node_list, cens, risk, id,
       outcome_type <- "continuous"
     }
 
-    learners <- check_variation(natural$train[i & rt, ][[outcome]], learners)
-
-    fit <- run_ensemble(natural$train[i & rt, c(id, vars, outcome)],
+    fit <- run_ensemble(natural$train[i & rt, c("lmtp_id", vars, outcome)],
                         outcome,
                         learners,
                         outcome_type,
-                        id,
+                        "lmtp_id",
                         control$.learners_outcome_folds)
 
     if (control$.return_full_fits) {
@@ -71,29 +63,27 @@ estimate_tmle <- function(natural, shifted, outcome, node_list, cens, risk, id,
       fits[[t]] <- extract_sl_weights(fit)
     }
 
-    trt_var <- names(shifted$train)[t]
-    under_shift_train <- natural$train[jt & rt, vars]
-    under_shift_train[[trt_var]] <- shifted$train[jt & rt, trt_var]
-
-    under_shift_valid <- natural$valid[jv & rv, vars]
-    under_shift_valid[[trt_var]] <- shifted$valid[jv & rv, trt_var]
-
-    m_natural_train[jt & rt, t] <- bound(SL_predict(fit, natural$train[jt & rt, vars]), 1e-05)
-    m_shifted_train[jt & rt, t] <- bound(SL_predict(fit, under_shift_train), 1e-05)
-    m_natural_valid[jv & rv, t] <- bound(SL_predict(fit, natural$valid[jv & rv, vars]), 1e-05)
-    m_shifted_valid[jv & rv, t] <- bound(SL_predict(fit, under_shift_valid), 1e-05)
-
-    wts <- {
-      if (is.null(weights))
-        ratios[i & rt, t]
-      else
-        ratios[i & rt, t] * weights[i & rt]
+    if (length(trt) > 1) {
+      trt_t <- trt[[t]]
+    } else {
+      trt_t <- trt[[1]]
     }
+
+    under_shift_train <- natural$train[jt & rt, c("lmtp_id", vars)]
+    under_shift_train[, trt_t] <- shifted$train[jt & rt, trt_t]
+
+    under_shift_valid <- natural$valid[jv & rv, c("lmtp_id", vars)]
+    under_shift_valid[, trt_t] <- shifted$valid[jv & rv, trt_t]
+
+    m_natural_train[jt & rt, t] <- bound(SL_predict(fit, natural$train[jt & rt, c("lmtp_id", vars)]), 1e-05)
+    m_shifted_train[jt & rt, t] <- bound(SL_predict(fit, under_shift_train), 1e-05)
+    m_natural_valid[jv & rv, t] <- bound(SL_predict(fit, natural$valid[jv & rv, c("lmtp_id", vars)]), 1e-05)
+    m_shifted_valid[jv & rv, t] <- bound(SL_predict(fit, under_shift_valid), 1e-05)
 
     fit <- sw(
       glm(
         natural$train[i & rt, ][[outcome]] ~ offset(qlogis(m_natural_train[i & rt, t])),
-        weights = wts,
+        weights = ratios[i & rt, t],
         family = "binomial"
       )
     )