+++ tests/testthat/test-point_treatment.R
@@ -13,13 +13,11 @@ Y <- Y.1 * A + Y.0 * (1 - A)
 tmp <- data.frame(W1, W2, A, Y, Y.1, Y.0)
 truth <- mean(tmp$Y.1)
 
-suppressWarnings({
-  sub <- lmtp_sub(tmp, "A", "Y", baseline = c("W1", "W2"), shift = static_binary_on, folds = 1)
-  ipw <- lmtp_ipw(tmp, "A", "Y", baseline = c("W1", "W2"), shift = static_binary_on, folds = 1)
-  tmle <- lmtp_tmle(tmp, "A", "Y", baseline = c("W1", "W2"), shift = static_binary_on, folds = 1)
-  tmle_riesz <- lmtp_tmle(tmp, "A", "Y", baseline = c("W1", "W2"), shift = static_binary_on, folds = 1, trt_method = "riesz", learners_trt = c("glm"))
-  sdr <- lmtp_sdr(tmp, "A", "Y", baseline = c("W1", "W2"), shift = static_binary_on, folds = 1)
-})
+sub <- sw(lmtp_sub(tmp, "A", "Y", baseline = c("W1", "W2"), shift = static_binary_on, folds = 1))
+ipw <- sw(lmtp_ipw(tmp, "A", "Y", baseline = c("W1", "W2"), shift = static_binary_on, folds = 1))
+tmle <- sw(lmtp_tmle(tmp, "A", "Y", baseline = c("W1", "W2"), shift = static_binary_on, folds = 1))
+tmle_riesz <- sw(lmtp_tmle(tmp, "A", "Y", baseline = c("W1", "W2"), shift = static_binary_on, folds = 1, trt_method = "riesz", learners_trt = c("glm")))
+sdr <- sw(lmtp_sdr(tmp, "A", "Y", baseline = c("W1", "W2"), shift = static_binary_on, folds = 1))
 
 # tests
 test_that("point treatment fidelity", {